refactor add_* methods
